# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Library Manageemnt API",
        "description": "Library Manageemnt API for Swagger integration",
        "contact": {
            "name": "Rajan Bhatta (+88 0162119979, aobak63@gmail.com)",
            "email": "aobak63@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/pages": {
            "get": {
                "tags": [
                    "Pages"
                ],
                "summary": "Get all pages",
                "operationId": "cd59deb18bfa6add22005d63640ccd06",
                "responses": {
                    "200": {
                        "description": "List of pages"
                    }
                },
                "security": [
                    {
                        "ApiTokenAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pages"
                ],
                "summary": "Create a new page",
                "operationId": "22ffe9506aba1ed019b4d54f3d7764ea",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "page_number",
                                    "content",
                                    "chapter_id"
                                ],
                                "properties": {
                                    "page_number": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Page content goes here..."
                                    },
                                    "chapter_id": {
                                        "type": "integer",
                                        "example": 4
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Page created"
                    }
                },
                "security": [
                    {
                        "ApiTokenAuth": []
                    }
                ]
            }
        },
        "/api/pages/{id}": {
            "get": {
                "tags": [
                    "Pages"
                ],
                "summary": "Get a page by ID",
                "operationId": "e1e47ccc00c6b4d242e9370dac64dd5d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Page details"
                    }
                },
                "security": [
                    {
                        "ApiTokenAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Pages"
                ],
                "summary": "Update a page",
                "operationId": "0db0c90b61af20d0c9eeb5f440b91081",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "page_number": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Updated content"
                                    },
                                    "chapter_id": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Page updated"
                    }
                },
                "security": [
                    {
                        "ApiTokenAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pages"
                ],
                "summary": "Delete a page",
                "operationId": "1d34244ac8efd74dd341c8f746a4ce30",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Page deleted"
                    }
                },
                "security": [
                    {
                        "ApiTokenAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "ApiTokenAuth": {
                "type": "apiKey",
                "description": "Token based authentication using X-API-TOKEN header",
                "name": "X-API-TOKEN",
                "in": "header"
            }
        }
    },
...
}